refactor(search): rename search context file
