+++ src/pages/search/index.tsx
@@ -2,7 +2,7 @@ import { Flex } from '@vtex/brand-ui'
 import styles from 'styles/search-page'
 import SearchSections from 'components/search-sections'
 import Sidebar from 'components/sidebar'
-import SearchContextProvider from 'utils/contexts/searchContext'
+import SearchContextProvider from 'utils/contexts/search'
 import SearchResults from 'components/search-results'
 
 const SearchPage = () => {
